Refactored sidebar profile component
