+++ components/common/PageLayout/components/Sidebar/SidebarSubmenu.tsx
@@ -1,33 +1,19 @@
-import { useTheme } from '@emotion/react';
 import styled from '@emotion/styled';
-import AddIcon from '@mui/icons-material/Add';
 import KeyboardArrowDownIcon from '@mui/icons-material/KeyboardArrowDown';
 import MenuOpenIcon from '@mui/icons-material/MenuOpen';
 import Box from '@mui/material/Box';
 import Button from '@mui/material/Button';
-import Divider from '@mui/material/Divider';
 import IconButton from '@mui/material/IconButton';
-import Menu from '@mui/material/Menu';
-import MenuItem from '@mui/material/MenuItem';
 import Tooltip from '@mui/material/Tooltip';
 import Typography from '@mui/material/Typography';
-import useMediaQuery from '@mui/material/useMediaQuery';
-import { bindTrigger, bindMenu } from 'material-ui-popup-state';
+import { bindTrigger } from 'material-ui-popup-state';
 import { usePopupState } from 'material-ui-popup-state/hooks';
-import NextLink from 'next/link';
-import { useCallback, useState } from 'react';
 
-import { CreateSpaceForm } from 'components/common/CreateSpaceForm';
-import { Modal } from 'components/common/Modal';
-import UserDisplay from 'components/common/UserDisplay';
-import { useUserDetails } from 'components/profile/components/UserDetails/hooks/useUserDetails';
 import { useCurrentSpace } from 'hooks/useCurrentSpace';
-import { useSpaces } from 'hooks/useSpaces';
-import { useUser } from 'hooks/useUser';
 
 import { headerHeight } from '../Header';
 
-import SpaceListItem from './SpaceListItem';
+import { SidebarProfile } from './SidebarProfile';
 import WorkspaceAvatar from './WorkspaceAvatar';
 
 const StyledButton = styled(Button)(
@@ -60,47 +46,10 @@ const SidebarHeader = styled(Box)(
   }`
 );
 
-export default function SidebarSubmenu({
-  closeSidebar,
-  logoutCurrentUser
-}: {
-  closeSidebar: () => void;
-  logoutCurrentUser: () => void;
-}) {
-  const theme = useTheme();
-  const showMobileFullWidthModal = !useMediaQuery(theme.breakpoints.down('sm'));
-
+export default function SidebarSubmenu({ closeSidebar }: { closeSidebar: () => void }) {
   const currentSpace = useCurrentSpace();
-  const { spaces, createNewSpace, isCreatingSpace, setSpaces, isLoaded } = useSpaces();
-  const [spaceFormOpen, setSpaceFormOpen] = useState(false);
-  const { user } = useUser();
-  const { handleUserUpdate, isSaving } = useUserDetails({
-    user: user!
-  });
-
-  function showSpaceForm() {
-    setSpaceFormOpen(true);
-  }
-
-  function closeSpaceForm() {
-    setSpaceFormOpen(false);
-  }
 
   const menuPopupState = usePopupState({ variant: 'popover', popupId: 'profile-dropdown' });
-
-  const changeOrderHandler = useCallback(
-    async (draggedProperty: string, droppedOnProperty: string) => {
-      const newOrder = spaces.map((s) => s.id);
-      const propIndex = newOrder.indexOf(draggedProperty); // find the property that was dragged
-      newOrder.splice(propIndex, 1); // remove the dragged property from the array
-      const droppedOnIndex = newOrder.indexOf(droppedOnProperty); // find the index of the space that was dropped on
-      newOrder.splice(droppedOnIndex, 0, draggedProperty); // add the property to the new index
-      await handleUserUpdate({ spacesOrder: newOrder });
-      const newOrderedSpaces = spaces.sort((a, b) => newOrder.indexOf(a.id) - newOrder.indexOf(b.id));
-      setSpaces(newOrderedSpaces);
-    },
-    [handleUserUpdate, spaces]
-  );
 
   return (
     <SidebarHeader className='sidebar-header' position='relative'>
@@ -117,59 +66,12 @@ export default function SidebarSubmenu({
           {currentSpace?.name}
         </Typography>
       </StyledButton>
-      <Menu onClick={menuPopupState.close} {...bindMenu(menuPopupState)} sx={{ maxWidth: '330px' }}>
-        <MenuItem
-          component={NextLink}
-          href='/nexus'
-          sx={{
-            display: 'grid',
-            gridTemplateColumns: 'auto 1fr',
-            gridTemplateRows: 'auto auto',
-            columnGap: 1
-          }}
-        >
-          <UserDisplay user={user} hideName gridColumn='1' gridRow='1/3' />
-          <Typography variant='body2' noWrap>
-            {user?.username}
-          </Typography>
-          <Typography variant='body2' color='secondary'>
-            My Profile
-          </Typography>
-        </MenuItem>
-        <Divider />
-        <Typography component='p' variant='caption' mx={2} mb={0.5}>
-          My Spaces
-        </Typography>
-        {spaces.map((_space) => (
-          <SpaceListItem
-            key={_space.id}
-            disabled={isSaving || !isLoaded || isCreatingSpace}
-            selected={currentSpace?.domain === _space.domain}
-            space={_space}
-            changeOrderHandler={changeOrderHandler}
-          />
-        ))}
-        <MenuItem onClick={showSpaceForm} data-test='spaces-menu-add-new-space'>
-          <AddIcon sx={{ m: '5px 15px 5px 8px' }} />
-          Create or join a space
-        </MenuItem>
-        <Divider />
-        <MenuItem onClick={logoutCurrentUser}>Sign out</MenuItem>
-      </Menu>
+      <SidebarProfile menuPopupState={menuPopupState} />
       <Tooltip title='Close sidebar' placement='bottom'>
         <IconButton onClick={closeSidebar} size='small' sx={{ position: 'absolute', right: 0, top: 12 }}>
           <MenuOpenIcon />
         </IconButton>
       </Tooltip>
-      <Modal
-        size='medium'
-        open={spaceFormOpen}
-        sx={{ width: showMobileFullWidthModal ? '100%' : undefined }}
-        onClose={closeSpaceForm}
-        mobileDialog
-      >
-        <CreateSpaceForm onCancel={closeSpaceForm} />
-      </Modal>
     </SidebarHeader>
   );
 }
